rewriting code to include prefabs and particlesystems
i think the code i made prior to chapman's changes got lost because of the file corruption, so in the meantime i worked on rewriting the code again along with making this branch to solve the file corruptions if they're still happening.
