+++ Mole Game - Brackeys Game Jam 2024/Assets/Prefabs/Confetti Particles.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 2a2c5f621155316489f0516fdba7832a
-NativeFormatImporter:
+guid: aad8136467569284c8d5fc2a58b54981
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 6200000
   userData: 
   assetBundleName: 
   assetBundleVariant:
