+++ Mole Game - Brackeys Game Jam 2024/Assets/Prefabs/Mole.prefab
@@ -61,7 +61,7 @@ Rigidbody2D:
   m_Interpolate: 0
   m_SleepingMode: 1
   m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Constraints: 4
 --- !u!61 &8441254980778024424
 BoxCollider2D:
   m_ObjectHideFlags: 0
