--- Mole Game - Brackeys Game Jam 2024/Assets/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f7f2a55a0dc9b9324824f2babc3b7c20
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
