--- Mole Game - Brackeys Game Jam 2024/Assets/Resources/RainDropPhysicsMaterial.physicsMaterial2D
@@ -1,11 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!62 &6200000
-PhysicsMaterial2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: RainDropPhysicsMaterial
-  friction: 0
-  bounciness: 0.25
