--- Mole Game - Brackeys Game Jam 2024/Assets/Scripts/HandleDirtCollision.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class HandleDirtCollision : MonoBehaviour
-{
-    void OnTriggerEnter2D(Collider2D collider)
-    {
-        GameObject rainDrop = collider.gameObject;
-        Rigidbody2D body = rainDrop.GetComponent<Rigidbody2D>();
-        body.velocity *= 0.1f;
-    }
-}
