--- Mole Game - Brackeys Game Jam 2024/Assets/Scripts/HandleRainCollision.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class HandleRainCollision : MonoBehaviour
-{
-    void OnCollisionEnter2D(Collision2D collision)
-    {
-        GameObject go = collision.gameObject;
-        if (go.name == "Rain Drop")
-        {
-            Debug.Log("Mole hit water!");
-            tag = "dead";
-        }
-    }
-}
