+++ Mole Game - Brackeys Game Jam 2024/ProjectSettings/TagManager.asset
@@ -3,7 +3,8 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags: []
+  tags:
+  - dead
   layers:
   - Default
   - TransparentFX
